+++ bin/test.sh
@@ -20,6 +20,7 @@ jsonapitest \
 	$SCRIPT_DIR/../test/jobs.json \
 	$SCRIPT_DIR/../test/models.json \
 	$SCRIPT_DIR/../test/users.json \
+	$SCRIPT_DIR/../test/audits.json \
 	$SCRIPT_DIR/../test/cleaning.json
 
 suite="$(jq -cr '.[0].suite' $SCRIPT_DIR/../jsonapitest-results.json)"