+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Types of changes
 ## 2023
 ### Milestone 26 - 2023-
 **Added**
+  - [x] Adding new route to get Audit Logs with some parameters to filter results
   - [x] Allows to set customRules on datapoints
   - [x] Additional Rule Engine eventType: replaceWithDistance. This event will get the distance in meter between an Object and the longitude-latitude passed along the datapoint payload
 
@@ -26,7 +27,7 @@ Types of changes
   - [x] 
 
 **Changed**
-  - [x] 
+  - [x] Remove BCC on /mail/changePassword
 
 ### Milestone 25 - 2023-07
 **Added**